+++ CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ### Fixed
 
+- Fix IMPORTANT! verify expiry and scopes checks on jwt tokens
 - Fix finding consumer by custom id
 - Fix runtime error on unexpected function, kong.log.error -> kong.log.err
+
 
 ## v1.0.0